+++ lib/exi18n.ex
@@ -18,8 +18,8 @@ defmodule ExI18n do
   @doc "Path to directory that contains all files with translations."
   def path, do: Application.get_env(:exi18n, :path) || "priv/locales"
 
-  @doc "Storage type used to store translations."
-  def storage, do: Application.get_env(:exi18n, :storage) || :yml
+  @doc "Loader type used to store translations."
+  def loader, do: Application.get_env(:exi18n, :loader) || :yml
 
   @doc """
   Search for translation in given `locale` based on provided `key`.